requirements.txt
# Core packages
flask>=2.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0

# AI and ML
openai>=1.0.0
google-genai>=0.2.0
sentence-transformers>=2.2.0

# Database
chromadb>=0.4.0
redis>=5.0.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Telegram
python-telegram-bot>=20.0

# Utilities
python-dateutil>=2.8.0
typing-extensions>=4.5.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Security
cryptography>=41.0.0
